Support networkFeatures attribute for Azure source
Supporting this attribute required adding support for Table attribute
types along with boolean attributes. The support for booleans was needed
because we choose to implement the networkFeatures attribute as a set -
which in Lua is done by createing a table where the indices are the
values (in this case, strings representing network features) and the
values are all booleans with the True value.

As part of this, we refactored the code the create values from
MetadataValue enums to send to Lua into its own function,
build_lua_value(), which makes the coder overall cleaner and more
flexible if we need to add more attribute types in the future.