Require a filename for cloud merge and cloud read subcommands
Previously, if no filename was specified, we read from STDIN. This could
create some confusion if the filename wasn't specified by mistake
resulting in the command hanging forever waiting for STDIN to supply the
ip ranges. Now, we require that a filename of "-" be specified if the
user intends for the ip ranges to be read from STDIN. If the user
actually wants to read the ranges from a file named "-" (which seems
unlikely), they can specify the file as "./-".