Create a workflow that tries to ping the cloud services
The goal is to test if something has changed on the cloud services side.
If that is the case, it would be good to detect it as soon as possible.
We also run these tests on all new pull requests. We do not block
merging on these tests, however, since otherwise we would be unable to
merge if any source went down.