Create Github Actions workflow to build precompiled binaries
We're building for Linux, Mac OS, and Windows (32 and 64 bit) as these
seems like the most likley targets that people might want that we can
build for. I'm not quite clear if its possible to build for M2 Mac OS -
but that might be an interesting thing to look into next.

We do the build when a new tag is created and then upload them to a
release for that tag, creating it if necessary.